ui/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
boto3==1.34.69
psycopg2-binary==2.9.9
